super_store_app/requirements.txt
elasticsearch==7.12.0
kafka-python==2.0.2
requests-aws4auth==1.0.1
aioboto3==9.6.0
aiokafka==0.7.2
orjson==3.8.3
JPype1==1.4.1
bson==0.5.10
python-gnupg==0.5.3
